# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ORVELL PULSE Wholesale ERP Backend API",
        "description": "Production REST API documentation for ORVELL PULSE Wholesale ERP system with Spatie RBAC, multi-company isolation, inventory tracking, cashier settlement, and automated reporting.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default API Host"
        }
    ],
    "paths": {
...
        "/api/whatsapp/webhook": {
            "post": {
                "tags": [
                    "WhatsApp Webhooks"
                ],
                "summary": "Inbound WhatsApp message webhook (test/direct shape)",
                "description": "Simple/test webhook: explicit agent_type. Reply returned in HTTP response.",
                "operationId": "780666a983a0ea91b7410ef1157c13cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "agent_type",
                                    "from",
                                    "message"
                                ],
                                "properties": {
                                    "agent_type": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "customer"
                                        ],
                                        "example": "customer"
                                    },
                                    "from": {
                                        "type": "string",
                                        "example": "+233240000001"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "catalog"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook processed and reply returned"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/whatsapp/chatterly": {
            "post": {
                "tags": [
                    "WhatsApp Webhooks"
                ],
                "summary": "Chatterly / Heywave inbound gateway webhook",
                "description": "Real Chatterly gateway webhook: resolve agent from instance, process, send reply back.",
                "operationId": "5020d0aac739c3e38e00fb1f1fc54022",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "instance": {
                                        "type": "string",
                                        "example": "orvell_accra_bot"
                                    },
                                    "instance_id": {
                                        "type": "string",
                                        "example": "inst_12345"
                                    },
                                    "data": {
                                        "properties": {
                                            "from": {
                                                "type": "string",
                                                "example": "+233240000001"
                                            },
                                            "body": {
                                                "type": "string",
                                                "example": "track ORD-001"
                                            },
                                            "fromMe": {
                                                "type": "boolean",
                                                "example": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook acknowledged and message queued"
                    }
                }
            }
        },
...
        "/api/orders/create": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create new sale / order",
                "operationId": "416b3022d3362001507184c5409f6956",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "buyer_id",
                                    "items"
                                ],
                                "properties": {
                                    "buyer_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "item_category_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/list": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders",
                "operationId": "b7302babcce43d8766db4246b240bac3",
                "responses": {
                    "200": {
                        "description": "Orders fetched successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/show/{uuid}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order by UUID",
                "operationId": "656b7b208e09ac2a50878132505fe55f",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Order UUID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order fetched successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter Sanctum personal access token as: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}